% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load.R
\name{are_df_names_valid}
\alias{are_df_names_valid}
\title{are_df_names_valid determines whether df has appropriate names. I.e., no spaces.}
\usage{
are_df_names_valid(col_names)
}
\arguments{
\item{col_names}{}
}
\value{
A boolean indicating whether the names are valid.
}
\description{
are_df_names_valid determines whether df has appropriate names. I.e., no spaces.
}
