% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_feature.R
\name{bake_recipe}
\alias{bake_recipe}
\title{Bake a recipe}
\usage{
bake_recipe(data, dep_var, ind_vars, vars_center)
}
\arguments{
\item{data}{dataframe containing the data (r$data)}

\item{dep_var}{the dependant variable}

\item{ind_vars}{the independant variables}

\item{vars_center}{list of variables for center transformation}
}
\value{
a baked data frame
}
\description{
Bake a recipe
}
