% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{df_to_dygraph}
\alias{df_to_dygraph}
\title{Functions which help with plotting}
\usage{
df_to_dygraph(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
dygraph to be rendered.
}
\description{
Functions which help with plotting
}
