% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_xts.R
\name{df_to_ts}
\alias{df_to_ts}
\title{Convert a dataframe to an time series object. Only works for Year Quarters.}
\usage{
df_to_ts(df, selected_vars = NULL, start = NULL, end = NULL)
}
\arguments{
\item{df}{}
}
\value{
an ts object generated from the dataframe.
}
\description{
Convert a dataframe to an time series object. Only works for Year Quarters.
}
