% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_xts.R
\name{df_to_xts}
\alias{df_to_xts}
\title{Convert a dataframe to an XTS object. Only works for Year Quarters}
\usage{
df_to_xts(df)
}
\arguments{
\item{df}{}
}
\value{
an xts object generated from the dataframe.
}
\description{
Convert a dataframe to an XTS object. Only works for Year Quarters
}
