% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_xts.R
\name{diff_inv_df}
\alias{diff_inv_df}
\title{diff_inv_df
Undo differencing.}
\usage{
diff_inv_df(df, starting_values)
}
\arguments{
\item{df}{- dataframe of differences}

\item{starting_values}{- dataframe of starting values (one row, one per
column in df).}
}
\value{
a dataframe
}
\description{
diff_inv_df
Undo differencing.
}
