% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model.R
\name{fit_decomp}
\alias{fit_decomp}
\title{fit_decomp - takes a data frame and returns the fit object for a time series
decomposition forecast on the data.}
\usage{
fit_decomp(df, dep_var, start = NULL, end = NULL)
}
\arguments{
\item{df}{data frame of data}

\item{dep_var}{string - name of dependent variable to forecast}

\item{start}{vector - start year and quarter in format c(YYYY, Q)}

\item{end}{vector - end year and quarter in format c(YYYY, Q)}
}
\value{
fit object for a time series decomposition forecast on the data in df
}
\description{
fit_decomp - takes a data frame and returns the fit object for a time series
decomposition forecast on the data.
}
