% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model.R
\name{fit_holtwinters}
\alias{fit_holtwinters}
\title{fit_holtwinters - takes a data frame and returns the fit object for a
Holt-Winters forecast on the data.}
\usage{
fit_holtwinters(df, dep_var, start, end)
}
\arguments{
\item{df}{data frame of data}

\item{dep_var}{string - name of dependent variable to forecast}

\item{start}{vector - start year and quarter in format c(YYYY, Q)}

\item{end}{vector - end year and quarter in format c(YYYY, Q)}
}
\value{
fit object for a Holt-Winters forecast on the data in df
}
\description{
fit_holtwinters - takes a data frame and returns the fit object for a
Holt-Winters forecast on the data.
}
