% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model.R
\name{fit_linear}
\alias{fit_linear}
\title{fit_linear - takes a data frame and returns the fit object for a linear
regression forecast on the data.}
\usage{
fit_linear(df, dep_var, ind_var, start, end, model)
}
\arguments{
\item{df}{data frame of data}

\item{dep_var}{string - name of dependent variable to forecast}

\item{ind_var}{string - vector of names of independent variables to fit}

\item{start}{vector - start year and quarter in format c(YYYY, Q)}

\item{end}{vector - end year and quarter in format c(YYYY, Q)}

\item{model}{string - the linear model e.g. "dep_var ~ ind1 + ind2 + ..."}
}
\value{
fit object for a linear regression forecast on the data in df
}
\description{
fit_linear - takes a data frame and returns the fit object for a linear
regression forecast on the data.
}
