% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model.R
\name{fit_models}
\alias{fit_models}
\title{fit_models - find model fit objects for all selected model types}
\usage{
fit_models(df, dep_var, ind_var, start, end, forecasts_to_include)
}
\arguments{
\item{df}{data frame containing data}

\item{dep_var}{string - name of dependent variable to forecast}

\item{ind_var}{string - vector of names of independent variables to fit}

\item{start}{vector - start year and quarter in format c(YYYY, Q)}

\item{end}{vector - end year and quarter in format c(YYYY, Q)}

\item{forecasts}{list of string values detailing which forecasts to include}
}
\value{
list of model fit objects
}
\description{
fit_models - find model fit objects for all selected model types
}
