% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model.R
\name{get_forecast_plotdata}
\alias{get_forecast_plotdata}
\title{get_forecast_plotdata}
\usage{
get_forecast_plotdata(fit, proj_data = NULL)
}
\arguments{
\item{fit}{- model object}

\item{proj_data}{- data frame of projected independent variable data to use
in forecast}
}
\value{
plotly line graph of forecast
}
\description{
A function that takes the model fit object and returns a data
frame containing the x and y data required for a plot of the actuals,
forecast, and 95% confidence interval.
}
