% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_subset.R
\name{get_model_formula}
\alias{get_model_formula}
\title{Helper function for cross-validation; model generator}
\usage{
get_model_formula(id, object, outcome)
}
\arguments{
\item{model.formula}{A vector of length}

\item{data}{The dataframe containing the regression data}

\item{nvars}{The max number of variables to use in the best subset model}
}
\description{
Helper function for cross-validation; model generator
}
