% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model.R
\name{get_proj_data}
\alias{get_proj_data}
\title{get_proj_data}
\usage{
get_proj_data(df, end)
}
\arguments{
\item{df}{data frame of data}

\item{end}{vector - end year and quarter in format c(YYYY, Q)}
}
\value{
data frame of projection data for all columns
}
\description{
Takes a data frame of time series data and an end point and
returns a data frame with just the data from after that point. To be used
to get project data for independent variables from a data frame of combined
historical and projection data.
}
