% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model.R
\name{get_starting_values}
\alias{get_starting_values}
\title{undiff_ts}
\usage{
get_starting_values(flg_diff, dep_var, starting_values = NULL)
}
\arguments{
\item{df}{time series of data}

\item{start_value}{double - starting value of time series (to add differences
to)}
}
\value{
data frame of time series
}
\description{
Takes
}
