% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_subset.R
\name{ifelsefun}
\alias{ifelsefun}
\title{Helper function for writing signif. codes for rows in lm_summaries}
\usage{
ifelsefun(x)
}
\arguments{
\item{x}{Generic variable for current column}
}
\value{

}
\description{
Helper function for writing signif. codes for rows in lm_summaries
}
