% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load.R
\name{is_df_continuous}
\alias{is_df_continuous}
\title{is_df_continous determines whether there are any missing values in the middle of a dataframe column.}
\usage{
is_df_continuous(df)
}
\arguments{
\item{df}{}
}
\value{
A boolean indicting whether there is missing data.
}
\description{
is_df_continous determines whether there are any missing values in the middle of a dataframe column.
}
