% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load.R
\name{is_row_index_unique}
\alias{is_row_index_unique}
\title{Checks whether dates are unique or if rows have multiple values.}
\usage{
is_row_index_unique(df)
}
\arguments{
\item{df}{}
}
\value{
A boolean indicating whether the rows are unique
}
\description{
Checks whether dates are unique or if rows have multiple values.
}
