% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load.R
\name{is_valid_df}
\alias{is_valid_df}
\title{is_valid_df - Determines whether a dataframe is valid to be used for forecasting.}
\usage{
is_valid_df(df)
}
\arguments{
\item{df}{a single dataframe to be checked.}
}
\value{
a boolean describing whether the dataframe is valid.
}
\description{
is_valid_df - Determines whether a dataframe is valid to be used for forecasting.
}
