% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load.R
\name{load_excel}
\alias{load_excel}
\title{load_excel}
\usage{
load_excel(filename)
}
\arguments{
\item{filename}{a string containing the filename.}
}
\value{
a dataframe containg the data.
}
\description{
load_excel
}
