% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load.R
\name{load_r_file}
\alias{load_r_file}
\title{read an r script}
\usage{
load_r_file(filename)
}
\arguments{
\item{filename}{a string of the r file to be run.}
}
\value{
a dataframe containg the data.
}
\description{
read an r script
}
