% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load.R
\name{load_rds_file}
\alias{load_rds_file}
\title{read a dataframe from an rds file}
\usage{
load_rds_file(filename)
}
\arguments{
\item{filename}{}
}
\value{
a dataframe
}
\description{
read a dataframe from an rds file
}
