% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load.R
\name{load_user_data}
\alias{load_user_data}
\title{Load and convert user selected data}
\usage{
load_user_data(upload)
}
\arguments{
\item{upload}{a dataframe containing a details of uploaded files.}
}
\value{
a dataframe containing the combines files
}
\description{
Load and convert user selected data
}
