% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load.R
\name{merge_user_data}
\alias{merge_user_data}
\title{Merge a list of valid dataframes into a single dataframe.}
\usage{
merge_user_data(df_list, cols = c("Year", "Quarter"))
}
\arguments{
\item{df_list}{A list of valid dataframes}

\item{cols}{The columns used to perform the join by. Default = c("Year", "Quarter")}
}
\value{
The merged dataframes as a single dataframe.
}
\description{
Merge a list of valid dataframes into a single dataframe.
}
