% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_best_subset.R
\name{mod_best_subset_ui}
\alias{mod_best_subset_ui}
\alias{mod_best_subset_server}
\title{mod_best_subset_ui and mod_best_subset_server}
\usage{
mod_best_subset_ui(id)

mod_best_subset_server(id, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
