% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_data.R
\name{mod_plot_data_ui}
\alias{mod_plot_data_ui}
\alias{mod_plot_data_server}
\title{mod_plot_data_ui and mod_plot_data_server}
\usage{
mod_plot_data_ui(id)

mod_plot_data_server(id, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
