% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_subset.R
\name{model.from.regobject}
\alias{model.from.regobject}
\title{Extracts lm and model string from regsubsets object}
\usage{
model.from.regobject(regobject, modelindex, modeldata)
}
\arguments{
\item{regobject}{regsubsets object}

\item{modelindex}{Index for model}

\item{modeldata}{Data within model}
}
\description{
Extracts lm and model string from regsubsets object
}
