% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_subset.R
\name{model_equation}
\alias{model_equation}
\title{Extracts an equation from the lm object}
\usage{
model_equation(model, ...)
}
\arguments{
\item{model}{The lm object for the model}
}
\value{

}
\description{
Extracts an equation from the lm object
}
