% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_subset_reg_attempt_golemified.R
\name{outputs}
\alias{outputs}
\title{Conduct best subset regression}
\usage{
outputs(dep.var, data, nvars)
}
\arguments{
\item{dep.var}{The dependent variable as a string}

\item{data}{The dataframe containing the regression data}

\item{nvars}{The number of variables to use in the best subset model}
}
\value{

}
\description{
Conduct best subset regression
}
