% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cormat.R
\name{plot_cormat}
\alias{plot_cormat}
\title{Generate the Collinearity Matrix plot for best_subset}
\usage{
plot_cormat(data, ind_vars)
}
\arguments{
\item{data}{dataframe containing the data}

\item{ind_vars}{character vector contain the names of indepenant variables}
}
\value{
ggploy object containing the collinearity matrix
}
\description{
Generate the Collinearity Matrix plot for best_subset
}
