% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model.R
\name{predict_models}
\alias{predict_models}
\title{predict_models - calculate forecasts}
\usage{
predict_models(model_fits, proj_indep_var, n_periods_to_forecast = 4)
}
\arguments{
\item{fit}{list - model object(s) containing fit}

\item{dep_var}{string - name of dependent variable to forecast}

\item{ind_var}{string - vector of names of independent variables to fit}

\item{start}{vector - start year and quarter in format c(YYYY, Q)}

\item{end}{vector - end year and quarter in format c(YYYY, Q)}
}
\value{
data frame containing predictions/forecasts for multiple models
}
\description{
predict_models - calculate forecasts
}
