% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{xts_to_df}
\alias{xts_to_df}
\title{Converts a xts object into a dataframe compatible with SHIFT}
\usage{
xts_to_df(x)
}
\arguments{
\item{x}{an xts object}
}
\value{
a dataframe
}
\description{
Converts a xts object into a dataframe compatible with SHIFT
}
