% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_subset.R
\name{allsubsetregression}
\alias{allsubsetregression}
\title{Conduct best subset regression
Conducts exhaustive best subset, cross-validation and stepwise search}
\usage{
allsubsetregression(dep_var, data, nvars)
}
\arguments{
\item{data}{The dataframe containing the regression data}

\item{nvars}{The number of variables to use in the best subset model}

\item{dep.var}{The dependent variable as a string}
}
\value{

}
\description{
Conduct best subset regression
Conducts exhaustive best subset, cross-validation and stepwise search
}
