% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_subset.R
\name{get_cv_error}
\alias{get_cv_error}
\title{Helper function for cross-validation; sampling part}
\usage{
get_cv_error(model.formula, data, nvars)
}
\arguments{
\item{model.formula}{The formula for the model}

\item{data}{The dataframe containing the regression data}

\item{nvars}{The max number of variables to use in the best subset model}
}
\value{

}
\description{
Helper function for cross-validation; sampling part
}
