% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_subset.R
\name{get_model_formula}
\alias{get_model_formula}
\title{Helper function for cross-validation; model generator}
\usage{
get_model_formula(id, models, dep_var)
}
\arguments{
\item{id}{The id for the model}

\item{models}{regsubsets object}

\item{dep_var}{The dependent variable as a string}
}
\description{
Helper function for cross-validation; model generator
}
