% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model.R
\name{plot_forecast}
\alias{plot_forecast}
\title{plot_forecast}
\usage{
plot_forecast(
  df,
  dep_var,
  ind_var = NULL,
  start,
  end,
  forecast_type,
  proj_data = NULL,
  diff_inv = FALSE,
  diff_starting_values = NULL
)
}
\arguments{
\item{df}{data frame of data}

\item{dep_var}{string - name of dependent variable to forecast}

\item{ind_var}{string - vector of names of indepedent variables to use in
forecast}

\item{start}{vector - start year and quarter in format c(YYYY, Q)}

\item{end}{vector - end year and quarter in format c(YYYY, Q)}

\item{proj_data}{data frame (default NULL) of projected data to use if
producing a linear regression forecast.}

\item{diff_inv}{boolean (default FALSE) - if data is differenced and you want
to undo this to compare to the original dataset set to TRUE}

\item{diff_starting_values}{numeric (default NULL) - if data is differenced,
where to start adding differences.}

\item{forecast_types}{- vector of string values giving forecast types to use.
One of: "holtwinters", "naive", "decomposition", "linear".}
}
\value{
plotly line graph of forecast
}
\description{
A function that takes a data frame and returns the plot of the
selected forecast for the selected column of data.
}
