% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model.R
\name{predict_models}
\alias{predict_models}
\title{predict_models - calculate forecasts}
\usage{
predict_models(model_fits, proj_data = NULL, n_periods_to_forecast = 4)
}
\arguments{
\item{model_fits}{list of model fit objects to use in forecast. Can use
fit_models() to calculate these.}

\item{proj_data}{if linear regression model is included in model_fits,
this data frame is required, containing projected data to use in the forecast}

\item{n_periods_to_forecast}{integer - number of periods to forecast forward.
Default is 4 periods.}
}
\value{
data frame containing predictions/forecasts for multiple models
}
\description{
A function that returns a list of forecasts for the input
model fits.
}
